requirements.txt
opencv_python==3.4.2.17
numpy==1.12.1
Django==2.1
Pillow==5.2.0
